# apps/inquaire-api-server/docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/inquiries": {
      "post": {
        "operationId": "InquiriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInquiryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "문의 생성",
        "tags": [
          "inquiries"
        ]
      },
      "get": {
        "operationId": "InquiriesController_findAll",
        "parameters": [
          {
            "name": "business_id",
            "required": false,
            "in": "query",
            "description": "사업체 ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel_id",
            "required": false,
            "in": "query",
            "description": "채널 ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customer_id",
            "required": false,
            "in": "query",
            "description": "고객 ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "상태",
            "schema": {
              "enum": [
                "NEW",
                "IN_PROGRESS",
                "COMPLETED",
                "ON_HOLD"
              ],
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "검색어 (메시지 텍스트)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_date",
            "required": false,
            "in": "query",
            "description": "시작 날짜 (ISO 8601)",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "end_date",
            "required": false,
            "in": "query",
            "description": "종료 날짜 (ISO 8601)",
            "schema": {
              "example": "2025-12-31",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호 (기본값: 1)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지 크기 (기본값: 20)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "정렬 필드 (기본값: received_at)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "정렬 방향 (asc/desc, 기본값: desc)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "문의 목록 조회",
        "tags": [
          "inquiries"
        ]
      }
    },
    "/api/v1/inquiries/stats": {
      "get": {
        "operationId": "InquiriesController_getStats",
        "parameters": [
          {
            "name": "business_id",
            "required": true,
            "in": "query",
            "description": "사업체 ID",
            "schema": {
              "example": "clxxx123456789",
              "type": "string"
            }
          },
          {
            "name": "start_date",
            "required": false,
            "in": "query",
            "description": "통계 시작 날짜 (ISO 8601 형식)",
            "schema": {
              "example": "2024-01-01",
              "type": "string"
            }
          },
          {
            "name": "end_date",
            "required": false,
            "in": "query",
            "description": "통계 종료 날짜 (ISO 8601 형식)",
            "schema": {
              "example": "2024-12-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "문의 통계 조회",
        "tags": [
          "inquiries"
        ]
      }
    },
    "/api/v1/inquiries/{id}": {
      "get": {
        "operationId": "InquiriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "문의 상세 조회",
        "tags": [
          "inquiries"
        ]
      },
      "patch": {
        "operationId": "InquiriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInquiryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "문의 업데이트",
        "tags": [
          "inquiries"
        ]
      },
      "delete": {
        "operationId": "InquiriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "문의 삭제",
        "tags": [
          "inquiries"
        ]
      }
    },
    "/api/v1/inquiries/{id}/analyze": {
      "post": {
        "operationId": "InquiriesController_analyzeWithAi",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "AI 분석 실행",
        "tags": [
          "inquiries"
        ]
      }
    },
...
  },
  "info": {
    "title": "AI 상담 매니저 API",
    "description": "\n## AI 상담 매니저 API 문서\n\nAI 기반 자동 문의 관리 시스템 REST API\n\n### 주요 기능\n- 🤖 AI 기반 문의 자동 분석 및 답변\n- 💬 Kakao Talk, LINE 메시징 플랫폼 연동\n- 📊 실시간 문의 통계 및 대시보드\n- 👥 고객 관리 및 이력 추적\n- 🏢 멀티 테넌시 (사업체별 격리)\n\n### 인증\n모든 API는 JWT Bearer 토큰을 사용합니다.\n\n```\nAuthorization: Bearer <access_token>\n```\n\n### 에러 응답\n모든 에러는 다음 형식을 따릅니다:\n\n```json\n{\n  \"statusCode\": 400,\n  \"message\": \"에러 메시지\",\n  \"timestamp\": \"2025-01-16T10:00:00.000Z\",\n  \"path\": \"/api/endpoint\"\n}\n```\n\n### Rate Limiting\n- 기본: 100 requests / 60 seconds\n- Auth 엔드포인트: 10 requests / 60 seconds\n  ",
    "version": "1.0.0",
    "contact": {
      "name": "AI Next Team",
      "url": "https://ainext.com",
      "email": "support@ainext.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    },
    {
      "url": "https://api.inquaire.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "JWT 액세스 토큰을 입력하세요",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateInquiryDto": {
        "type": "object",
        "properties": {
          "business_id": {
            "type": "string",
            "description": "사업체 ID",
            "example": "bus-a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "channel_id": {
            "type": "string",
            "description": "채널 ID",
            "example": "ch-12345678-90ab-cdef-1234-567890abcdef"
          },
          "customer_id": {
            "type": "string",
            "description": "고객 ID",
            "example": "cust-abcdef12-3456-7890-abcd-ef1234567890"
          },
          "message_text": {
            "type": "string",
            "description": "메시지 텍스트",
            "example": "안녕하세요, 예약 문의 드립니다. 이번 주 토요일 오전 10시 가능할까요?"
          },
          "platform_message_id": {
            "type": "string",
            "description": "플랫폼 메시지 ID",
            "example": "kakao_msg_123456789"
          },
          "extracted_info": {
            "type": "object",
            "description": "추출된 정보 (JSON)",
            "example": {
              "desired_date": "2025-11-23",
              "desired_time": "오전",
              "service_type": "예약"
            }
          }
        },
        "required": [
          "business_id",
          "channel_id",
          "customer_id",
          "message_text"
        ]
      },
      "UpdateInquiryDto": {
        "type": "object",
        "properties": {
          "reply_text": {
            "type": "string",
            "description": "답변 텍스트"
          },
          "status": {
            "type": "string",
            "description": "상태",
            "enum": [
              "NEW",
              "IN_PROGRESS",
              "COMPLETED",
              "ON_HOLD"
            ]
          },
          "summary": {
            "type": "string",
            "description": "요약"
          },
          "type": {
            "type": "string",
            "description": "문의 유형"
          },
          "notes": {
            "type": "string",
            "description": "관리자 메모"
          },
          "extracted_info": {
            "type": "object",
            "description": "추출된 정보 (JSON)"
          }
        }
      },
...
    }
  }
}